--- demooldhooks/views/templates/field.html.twig
@@ -1,2 +1,0 @@
-<h3>{{ field_name }}</h3>
-{{ form_widget(form) }}
