+++ demooldproductpagehooks/config.xml
