+++ demooldproductpagehooks/views/templates/field.html.twig
@@ -1,0 +1,10 @@
+<style>
+    .product-page-hook {
+        padding: 20px; margin: 5px;
+        border: 3px dotted green;
+    }
+</style>
+<div class="product-page-hook">
+    <h3>{{ field_name }}</h3>
+    {{ form_widget(form) }}
+</div>
